Completion Time

Checks when the specified circle was first entered.

Usage

time_circle_entered(
  x_vector,
  y_vector,
  t_vector,
  xMid = 0,
  yMid = 0,
  radius = 1,
  include_radius = TRUE
)

Arguments

x_vector

x-coordinates of the executed trajectory.

y_vector

y-coordinates of the executed trajectory.

t_vector

Timestamps of the executed trajectory.

xMid

x-coordinate of the center of the circle.

yMid

y-coordinate of the center of the circle.

radius

radius of the center of the circle.

include_radius

Whether points lying exactly on the radius should be included in the circle. Defaults to TRUE.

Value

Value of t_vector at the first time at which the trajectory is in the circle.

References

Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020). Design choices: Empirical recommendations for designing two-dimensional finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416. doi:10.3758/s13428-020-01409-0

Examples

time_circle_entered(0:10, rep(0,11), 0:10, xMid=10, yMid=0, radius=1)
#> [1] 9
time_circle_entered(0:10, rep(0,11), 0:10, xMid=10, yMid=0, radius=1, include_radius = FALSE)
#> [1] 10